Updated Gradle and Android plugin versions
This commit updates the Gradle wrapper to version 8.13 and the Android Gradle Plugin to version 8.11.1.

Additionally, a new run configuration and deployment target selector XML file were added to the `.idea` directory, and the `TrackScreen.kt` was modified to display track information (name, length, country, type) above the track drawing.